getUser API refactoring
