+++ src/app/api/user/route.ts
@@ -1,22 +1,12 @@
 import { UserProps, UserToken } from '@/app/interfaces/User';
-import {
-  USER_TABLE_NAME,
-  USER_TABLE_PARTITION_ID,
-} from '@/app/utilities/auth/constants';
 import isAuthenticated from '@/app/utilities/auth/isAuthenticated';
 import assertUserProps from '@/app/utilities/validation/assertUserData';
-import {
-  GetItemCommand,
-  GetItemCommandInput,
-  GetItemCommandOutput,
-  UpdateItemCommandOutput,
-} from '@aws-sdk/client-dynamodb';
-import { marshall, unmarshall } from '@aws-sdk/util-dynamodb';
+import { UpdateItemCommandOutput } from '@aws-sdk/client-dynamodb';
 import { NextRequest } from 'next/server';
-import { dbDocumentClient } from '../../utilities/db/configure';
 import updateUser from '@/app/utilities/auth/updateUser';
 import createUser from '@/app/utilities/auth/createUser';
 import { returnAuthError } from '@/app/utilities/auth/responses';
+import getUser from '@/app/utilities/auth/getUser';
 
 export async function GET(req: NextRequest) {
   try {
@@ -26,30 +16,16 @@ export async function GET(req: NextRequest) {
       return userResponse;
     }
 
-    const user = userResponse as UserToken;
-    const { email } = user;
+    const { email } = userResponse;
+    let user = await getUser(email);
 
-    const commandParams: GetItemCommandInput = {
-      TableName: USER_TABLE_NAME,
-      Key: marshall({
-        [USER_TABLE_PARTITION_ID]: email,
-      }),
-    };
-
-    const command = new GetItemCommand(commandParams);
-    const res: GetItemCommandOutput = await dbDocumentClient.send(command);
-
-    let { Item } = res;
-
-    if (!Item) {
+    if (!user) {
       // NOTE
       // This is possible when idP user signs up
-      Item = marshall(await createUser(email));
+      user = await createUser(email);
     }
 
-    const dbUser = unmarshall(Item);
-
-    return new Response(JSON.stringify(dbUser), {
+    return new Response(JSON.stringify(user), {
       status: 200,
       headers: new Headers({
         'Content-Type': 'application/json',
