+++ src/app/utilities/auth/createUser.ts
@@ -6,15 +6,11 @@ import {
 } from './constants';
 import { marshall } from '@aws-sdk/util-dynamodb';
 import { dbDocumentClient } from '../db/configure';
-import { UserProps } from '@/app/interfaces/User';
+import { UserWithEmailProps } from '@/app/interfaces/User';
 
-type NewUserProps = UserProps & {
-  Email: string;
-};
-
-const createUser = async (email: string): Promise<NewUserProps> => {
-  const defaultUser: Partial<NewUserProps> = { ...DEFAULT_USER };
-  let key: keyof NewUserProps;
+const createUser = async (email: string): Promise<UserWithEmailProps> => {
+  const defaultUser: Partial<UserWithEmailProps> = { ...DEFAULT_USER };
+  let key: keyof UserWithEmailProps;
   for (key in defaultUser) {
     const value = defaultUser[key];
     if (typeof value === 'number' || typeof value === 'boolean') {
@@ -26,7 +22,7 @@ const createUser = async (email: string): Promise<NewUserProps> => {
     }
   }
 
-  const user: NewUserProps = {
+  const user: UserWithEmailProps = {
     ...defaultUser,
     [USER_TABLE_PARTITION_ID]: email,
   };
